+++ cluster/pulumi/cluster/src/index.ts
@@ -1,0 +1,3 @@
+import { installNodePools } from './nodePools';
+
+installNodePools();